requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
celery==5.3.4
redis==5.0.0
pytest==7.4.0
SQLAlchemy==2.0.20
python-dotenv==1.0.0
psycopg2-binary==2.9.9